Move graph-layer-context and graph-model-context to own files.
