+++ src/plugins/graph/components/graph-axis.tsx
@@ -7,7 +7,7 @@ import {Active} from "@dnd-kit/core";
 import {useInstanceIdContext} from "../imports/hooks/use-instance-id-context";
 import {AttributeType} from "../../../models/data/attribute";
 import { IDataSet } from "../../../models/data/data-set";
-import {useGraphModelContext} from "../models/graph-model";
+import { useGraphModelContext } from "../hooks/use-graph-model-context";
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context";
 import {useGraphLayoutContext} from "../models/graph-layout";
 import {getDragAttributeInfo, useDropHandler} from "../imports/hooks/use-drag-drop";
