+++ src/plugins/graph/components/legend/layer-legend.tsx
@@ -6,7 +6,7 @@ import { GraphPlace } from "../../imports/components/axis-graph-shared";
 import { SimpleAttributeLabel } from "../simple-attribute-label";
 import { AddSeriesButton } from "./add-series-button";
 import { ReadOnlyContext } from "../../../../components/document/read-only-context";
-import { useGraphModelContext } from "../../models/graph-model";
+import { useGraphModelContext } from "../../hooks/use-graph-model-context";
 import { getSharedModelManager } from "../../../../models/tiles/tile-environment";
 import { isSharedDataSet, SharedDataSet, SharedDataSetType } from "../../../../models/shared/shared-data-set";
 import { useDataConfigurationContext } from "../../hooks/use-data-configuration-context";
