+++ src/plugins/graph/components/simple-attribute-label.tsx
@@ -5,7 +5,7 @@ import {GraphPlace } from "../imports/components/axis-graph-shared";
 import {AttributeType} from "../../../models/data/attribute";
 import {AxisOrLegendAttributeMenu} from "../imports/components/axis/components/axis-or-legend-attribute-menu";
 import { useDataConfigurationContext } from "../hooks/use-data-configuration-context";
-import { useGraphModelContext } from "../models/graph-model";
+import { useGraphModelContext } from "../hooks/use-graph-model-context";
 import { IDataSet } from "../../../models/data/data-set";
 import { kGraphClassSelector } from "../graph-types";
 import { ReadOnlyContext } from "../../../components/document/read-only-context";
