+++ src/plugins/graph/hooks/use-movables.ts
@@ -1,7 +1,7 @@
 import {useCallback, useEffect} from "react";
 import {ScaleNumericBaseType} from "../imports/components/axis/axis-types";
 import {useGraphLayoutContext} from "../models/graph-layout";
-import {useGraphModelContext} from "../models/graph-model";
+import { useGraphModelContext } from "../hooks/use-graph-model-context";
 import {onAnyAction} from "../../../utilities/mst-utils";
 import { IMovableLineModel } from "../adornments/movable-line/movable-line-model";
 import { IMovableValueModel } from "../adornments/movable-value/movable-value-model";
